Remove test data generator
