+++ apps/docs/content/components/contribution-graph.mdx
@@ -67,4 +67,3 @@ Customize individual blocks with className and style props:
 - The calendar will automatically fill gaps between dates with zero activity
 - You can control the calendar's date range by including empty entries as the first and last items
 - The component automatically handles dark mode when using the system color scheme
-- Use the `generateTestData()` helper function to quickly generate sample data for testing
