+++ apps/docs/examples/contribution-graph-minimal.tsx
@@ -2,12 +2,32 @@
 
 import {
   ContributionGraph,
-  ContributionGraphCalendar,
   ContributionGraphBlock,
-  generateTestData,
+  ContributionGraphCalendar,
 } from '@repo/contribution-graph';
+import { eachDayOfInterval, endOfYear, formatISO, startOfYear } from 'date-fns';
 
-const data = generateTestData();
+const maxCount = 20;
+const maxLevel = 4;
+const now = new Date();
+const days = eachDayOfInterval({
+  start: startOfYear(now),
+  end: endOfYear(now),
+});
+
+const data = days.map((date) => {
+  const c = Math.round(
+    Math.random() * maxCount - Math.random() * (0.8 * maxCount)
+  );
+  const count = Math.max(0, c);
+  const level = Math.ceil((count / maxCount) * maxLevel);
+
+  return {
+    date: formatISO(date, { representation: 'date' }),
+    count,
+    level,
+  };
+});
 
 const Example = () => (
   <ContributionGraph data={data}>
