+++ apps/docs/examples/contribution-graph-tooltip.tsx
@@ -2,11 +2,11 @@
 
 import {
   ContributionGraph,
-  ContributionGraphCalendar,
   ContributionGraphBlock,
+  ContributionGraphCalendar,
   ContributionGraphFooter,
-  generateTestData,
 } from '@repo/contribution-graph';
+import { eachDayOfInterval, endOfYear, formatISO, startOfYear } from 'date-fns';
 import {
   Tooltip,
   TooltipContent,
@@ -14,7 +14,27 @@ import {
   TooltipTrigger,
 } from '@/components/ui/tooltip';
 
-const data = generateTestData();
+const maxCount = 20;
+const maxLevel = 4;
+const now = new Date();
+const days = eachDayOfInterval({
+  start: startOfYear(now),
+  end: endOfYear(now),
+});
+
+const data = days.map((date) => {
+  const c = Math.round(
+    Math.random() * maxCount - Math.random() * (0.8 * maxCount)
+  );
+  const count = Math.max(0, c);
+  const level = Math.ceil((count / maxCount) * maxLevel);
+
+  return {
+    date: formatISO(date, { representation: 'date' }),
+    count,
+    level,
+  };
+});
 
 const Example = () => (
   <TooltipProvider>
@@ -26,9 +46,9 @@ const Example = () => (
               <g>
                 <ContributionGraphBlock
                   activity={activity}
+                  className="cursor-pointer"
                   dayIndex={dayIndex}
                   weekIndex={weekIndex}
-                  className="cursor-pointer"
                 />
               </g>
             </TooltipTrigger>
